0.99.53.1

+++ src/hlqest.cpp
@@ -43,7 +43,7 @@ int rng(int max) //range : [0, max]
 // [[Rcpp::export]]
 double hlqest(NumericVector x) {
   
-  int n = x.size();
+  long long n = x.size();
   
   std::sort(x.begin(), x.end(), [](double &a, double &b){ return a<b; });
   
@@ -55,26 +55,26 @@ double hlqest(NumericVector x) {
   }
   }
   double retval = 0.0;
-  int *lb = (int *)safe_malloc(sizeof(int) * n);
-  int *rb = (int *)safe_malloc(sizeof(int) * n);
-  int *q = (int *)safe_malloc(sizeof(int) * n);
+  long long *lb = (long long *)safe_malloc(sizeof(long long) * n);
+  long long *rb = (long long *)safe_malloc(sizeof(long long) * n);
+  long long *q = (long long *)safe_malloc(sizeof(long long) * n);
   
 // not needed (?):  rng(434977);  // set up random number generator
   
-  int nn = n * (n + 1) / 2;  // total number of pairs
-  int k1 = (nn + 1) / 2;     // median(s)
-  int k2 = (nn + 2) / 2;
+  long long nn = n * (n + 1) / 2;  // total number of pairs
+  long long k1 = (nn + 1) / 2;     // median(s)
+  long long k2 = (nn + 2) / 2;
   
-  for (int i = 0; i < n; i++) {  // probably used for indexing, therefore adjust to n-1
+  for (long long i = 0; i < n; i++) {  // probably used for indexing, therefore adjust to n-1
     lb[i] = i + 1;
     rb[i] = n;
   }
   
-  int sm = nn;  // number in set s at step m
-  int l = 0;    // number of pairs less than those in set s at step m
+  long long sm = nn;  // number in set s at step m
+  long long l = 0;    // number of pairs less than those in set s at step m
   
-  int mdll = (n + 1) / 2;
-  int mdlu = (n + 2) / 2;
+  long long mdll = (n + 1) / 2;
+  long long mdlu = (n + 2) / 2;
   
   double am = x[mdll - 1] + x[mdlu - 1];
   double amx, amn;
@@ -82,9 +82,9 @@ double hlqest(NumericVector x) {
   // q[i] = how many pairs x[i]+x[j] < am in row i
   
   for (;;) {  // 80
-    int j = n;
-    int sq = 0;
-    for (int i = 1; i <= n; i++)  // go until diagonal hit, or the pairs are no longer < am
+    long long j = n;
+    long long sq = 0;
+    for (long long i = 1; i <= n; i++)  // go until diagonal hit, or the pairs are no long longer < am
     {
       q[i - 1] = 0;
       for (; j >= i; j--) {
@@ -99,12 +99,12 @@ double hlqest(NumericVector x) {
       // 30
       amx = x[0] + x[0];
       amn = x[n - 1] + x[n - 1];
-      for (int i = 0; i < n; i++) {
+      for (long long i = 0; i < n; i++) {
         if (lb[i] > rb[i]) {  // skip this row if no element in it is in set s on this step
           continue;
         }
-        int lbi = lb[i];
-        int rbi = rb[i];
+        long long lbi = lb[i];
+        long long rbi = rb[i];
         
         double curi_x = x[i];
         amn = std::min(amn, x[lbi - 1] + curi_x);  // get the smallest in this row
@@ -128,9 +128,9 @@ double hlqest(NumericVector x) {
         // find: max of those < am, min of those >= am
         amn = x[n - 1] + x[n - 1];
         amx = x[0] + x[0];
-        for (int i = 1; i <= n; i++) {
-          int iq = q[i - 1];
-          int ipiq = i + iq;
+        for (long long i = 1; i <= n; i++) {
+          long long iq = q[i - 1];
+          long long ipiq = i + iq;
           double curi_x = x[i - 1];
           if (iq > 0) {
             amx = std::max(amx, curi_x + x[ipiq - 2]);
@@ -157,27 +157,27 @@ double hlqest(NumericVector x) {
       }
     if (sq < k1) {
       // 140
-      for (int j = 0; j < n; j++) {  // reset left bounds for each row
+      for (long long j = 0; j < n; j++) {  // reset left bounds for each row
         lb[j] = j + q[j] + 1;      // no subtraction in original 1-indexed fortran code.
       }
     } else {  // sq > k1 as sq != k1
       // 120
-      for (int j = 0; j < n; j++) {  // reset right bounds for each row
+      for (long long j = 0; j < n; j++) {  // reset right bounds for each row
         rb[j] = j + q[j];
       }
     }
     // 160
     l = 0;   // number of pairs less than those in new set s
     sm = 0;  // number of pairs still in new set s
-    for (int j = 1; j <= n; j++) {
+    for (long long j = 1; j <= n; j++) {
       l += lb[j - 1] - j;
       sm += rb[j - 1] - lb[j - 1] + 1;
     }
     // 170
     if (sm > 2) {
       // 50
-      int k = rng(sm);  // k is a random int from 0 to sm
-      for (int i = 0; i < n; i++) {
+      long long k = rng(sm);  // k is a random int from 0 to sm
+      for (long long i = 0; i < n; i++) {
         j = i + 1;
         if (k <= rb[i] - lb[i]) {
           break;
@@ -186,7 +186,7 @@ double hlqest(NumericVector x) {
       }
       // 70
       // j is a random row, get its median
-      int mdlrow = (lb[j - 1] + rb[j - 1]) / 2;
+      long long mdlrow = (lb[j - 1] + rb[j - 1]) / 2;
       am = x[j - 1] + x[mdlrow - 1];
       continue;  // go to 80
     } else {       // can only get to 2 left if k1 != k2, get their average
@@ -196,12 +196,12 @@ double hlqest(NumericVector x) {
       
       amx = x[0] + x[0];
       amn = x[n - 1] + x[n - 1];
-      for (int i = 0; i < n; i++) {
+      for (long long i = 0; i < n; i++) {
         if (lb[i] > rb[i]) {  // skip this row if no element in it is in set s on this step
           continue;
         }
-        int lbi = lb[i];
-        int rbi = rb[i];
+        long long lbi = lb[i];
+        long long rbi = rb[i];
         
         double curi_x = x[i];
         amn = std::min(amn, x[lbi - 1] + curi_x);  // get the smallest in this row
@@ -234,15 +234,15 @@ double hlqest(NumericVector x) {
 // [[Rcpp::export]]
 double hl2qest(NumericVector x, NumericVector y) {
   
-  int m = x.size();
-  int n = y.size();
+  long long m = x.size();
+  long long n = y.size();
   
   std::sort(x.begin(), x.end(), [](double &a, double &b){ return a<b; });
   std::sort(y.begin(), y.end(), [](double &a, double &b){ return a<b; });
   
-  int nn = m * n;
-  int k1 = (nn + 1) / 2;
-  int k2 = (nn + 2) / 2;
+  long long nn = m * n;
+  long long k1 = (nn + 1) / 2;
+  long long k2 = (nn + 2) / 2;
   
   if (n == 1) {
     return (x[k1 - 1] + x[k2 - 1]) / 2. - y[0];
@@ -250,24 +250,24 @@ double hl2qest(NumericVector x, NumericVector y) {
   if (m == 1) {
     return x[0] - (y[k1 - 1] + y[k2 - 1]) / 2.;
   }
-  int *lb = (int *)safe_malloc(sizeof(int) * n);
-  int *rb = (int *)safe_malloc(sizeof(int) * n);
-  int *q = (int *)safe_malloc(sizeof(int) * n);
+  long long *lb = (long long *)safe_malloc(sizeof(long long) * n);
+  long long *rb = (long long *)safe_malloc(sizeof(long long) * n);
+  long long *q = (long long *)safe_malloc(sizeof(long long) * n);
   
   double retval = 0.0;
   double am, amx, amn;
   
-  for (int i = 0; i < n; i++) {
-    lb[i] = std::max(1, (m + 1) - (nn + 1 - k1) / (i + 1));
+  for (long long i = 0; i < n; i++) {
+    lb[i] = std::max(1LL, (m + 1) - (nn + 1 - k1) / (i + 1));
     rb[i] = std::min(m, k2 / (n - i));
   }
   
-  int sm = nn;  // number in set s at step m
-  int l = 0;    // number of pairs less than those in set s at step m
-  int j = 0;
-  int sq = 0;
+  long long sm = nn;  // number in set s at step m
+  long long l = 0;    // number of pairs less than those in set s at step m
+  long long j = 0;
+  long long sq = 0;
   
-  int method = 1;
+  long long method = 1;
   
   for (;;) {
     switch (method) {
@@ -277,11 +277,11 @@ double hl2qest(NumericVector x, NumericVector y) {
     case 2:
       amx = x[0] - y[n - 1];
       amn = x[m - 1] - y[0];
-      for (int i = 0; i < n; i++) {
+      for (long long i = 0; i < n; i++) {
         if (lb[i] <= rb[i]) {  // skip this row if no element in it is in set s on this step
-          int lbi = lb[i] - 1;
+          long long lbi = lb[i] - 1;
           amn = std::min(amn, x[lbi] - y[i]);  // get the smallest in this row
-          int rbi = rb[i] - 1;
+          long long rbi = rb[i] - 1;
           amx = std::max(amx, x[rbi] - y[i]);  // get the largest in this row
         }
       }
@@ -295,23 +295,23 @@ double hl2qest(NumericVector x, NumericVector y) {
       }
       break;
     default:
-//      int k = std::rand() % sm;
-      int k = rng(sm);  // k is a random int from 0 to sm
+//      long long k = std::rand() % sm;
+      long long k = rng(sm);  // k is a random int from 0 to sm
 
-    for (int i = 0; i < n; i++) {
+    for (long long i = 0; i < n; i++) {
       j = i;
       if (k <= (rb[i] - lb[i]))
         break;
       k = k - rb[i] + lb[i] - 1;
     }
-    int mdlrow = (lb[j] + rb[j]) / 2 - 1;
+    long long mdlrow = (lb[j] + rb[j]) / 2 - 1;
     am = x[mdlrow] - y[j];
     break;
     }
     // Partition Step
     j = 0;
     sq = 0;
-    for (int i = 1; i <= n; i++) {
+    for (long long i = 1; i <= n; i++) {
       q[i - 1] = j;
       while (j < m) {
         if ((x[j] - y[i - 1]) >= am)
@@ -328,17 +328,17 @@ double hl2qest(NumericVector x, NumericVector y) {
       if ((sq == (k2 - 1)) || (sq == k1))
         break;
       if (sq > k1) {
-        for (int i = 0; i < n; i++) {
+        for (long long i = 0; i < n; i++) {
           rb[i] = q[i];
         }
       } else {
-        for (int i = 0; i < n; i++) {
+        for (long long i = 0; i < n; i++) {
           lb[i] = q[i] + 1;
         }
       }
       l = 0;
       sm = 0;
-      for (int i = 0; i < n; i++) {
+      for (long long i = 0; i < n; i++) {
         l += lb[i] - 1;
         sm += rb[i] - lb[i] + 1;
       }
@@ -350,8 +350,8 @@ double hl2qest(NumericVector x, NumericVector y) {
   
   amn = x[m - 1] - y[0];
   amx = x[0] - y[n - 1];
-  for (int i = 0; i < n; i++) {
-    int iq = q[i];
+  for (long long i = 0; i < n; i++) {
+    long long iq = q[i];
     if (iq > 0)
       amx = std::max(amx, x[iq - 1] - y[i]);
     if (iq < m)
